requirements.txt
ruamel.yaml==0.17.21
pydantic==1.10.7
tqdm==4.65.0